+++ src/main/java/me/Fupery/ArtMap/IO/MapId.java
@@ -1,15 +1,15 @@
 package me.Fupery.ArtMap.IO;
 
 public class MapId {
-    protected final short id;
+    protected final int id;
     protected final Integer hash;
 
-    public MapId(short id, int hash) {
+    public MapId(int id, int hash) {
         this.id = id;
         this.hash = hash;
     }
 
-    public short getId() {
+    public int getId() {
         return id;
     }
